Added Shadowflame item, Fishing weapon recipes
-Added SHADOWFLAME
This item is currently usable to convert Shadowflame weapons into
Shadowflames, which will eventually be used to craft the Shadowflame
weapons. (That is, once I learn how to get items to drop from existing
enemies)
-Added more frameworks to call from
-Rename to Namespace from MechParadise to TerrariaQuickRecipes-Mod,
which is the current file name due to the GitHib naming system
-Added recipes for Crystal Serpent, Toxikarp and Bladetongue using 20
of the biome’s common potion fish, 15 relevant materials (Crystal
Shard, Spider Fang, Ichor) and 10 souls of the correct infection type.