Build 0.1.1 - Martian Energy Core
Added MARTIAN ENERGY CORE
This ingredient is the Shadowflame equivalent for the Martian Madness,
but for now does not create any weapons, just like the other invasion
materials. You can break down all Martian Saucer drops (Except the
Anti-Gravity Hook and the Cosmic Car Key, which will not be craftable
either) into this item for now. In a more complete build these items
will be used in recipes.