0.2 Patch
-Added UNHALLOWED SHARD
This item is the Wall of Flesh bridging material. Currently this item
is created at a Mythril Anvil, but soon will be switched over to the
new Work station.
-Added ULTIMATE LOOT RECONSTRUCTOR
This work station is currently WIP, but will, when complete, be the way
to swap between different boss drops using the bridging material.
However, it is not an equivalent exchange, and secondary items may be
needed for crafting.
-Made all new items take up a seperate .cs file (The example mod does
this and I’d rather not take chances)
-Added a "Tiles" folder to the source for the Ultimate Boss
Reconstructor’s tile